feat: change state hoisting to external to be more controllable
